refactor(publikator): merge MemoContainer into MemoTree
